Link to PDF and display PNG in README.md
Ideally we would simply embed the PDF in the README.md and let Github
render that. Unfortunately Github still won't render embedded PDFs but I
still want a visitor on Github to immediately see the rendered CV.
To make it as easy as possible for a visitor to see the PDF which is
more interactive than an image, I embedded the PDF before the images.